Separating files
Separating the API class from blsPipeline file
